Get X11/OpenGL to render properly

+++ api/_patches/azul.rs/gl.rs
@@ -199,8 +199,8 @@
         pub type c_ushort = u16;
         pub type c_int = i32;
         pub type c_uint = u32;
-        pub type c_long = i32;
-        pub type c_ulong = u32;
+        pub type c_long = i64;
+        pub type c_ulong = u64;
         pub type c_longlong = i64;
         pub type c_ulonglong = u64;
         pub type c_float = f32;
